feat(rules): transition AI improvement to dedicated full-screen diff view
